+++ clients/spark/core/src/main/scala/io/treeverse/gc/DataLister.scala
@@ -11,9 +11,8 @@ import org.apache.hadoop.fs.FileStatus
 
 /** List all the files under a given path.
  */
-trait DataLister {
-  @transient lazy val spark = SparkSession.active
-
+abstract class DataLister {
+  @transient lazy val spark: SparkSession = SparkSession.active
   def listData(configMapper: ConfigMapper, path: Path): DataFrame
 }
 
@@ -25,7 +24,7 @@ class NaiveDataLister extends DataLister {
     val dataList = new ListBuffer[(String, Long)]()
     while (dataIt.hasNext) {
       val fileStatus = dataIt.next()
-      dataList += ((fileStatus.getPath.toString, fileStatus.getModificationTime()))
+      dataList += ((fileStatus.getPath.getName, fileStatus.getModificationTime))
     }
     dataList.toDF("address", "last_modified")
   }
@@ -36,10 +35,12 @@ class FileDescriptor(val path: String, val lastModified: Long) extends Serializa
 class ParallelDataLister extends DataLister with Serializable {
   def listPath(
       configMapper: ConfigMapper,
-      path: String
+      p: Path
   ): List[FileDescriptor] = {
-    val p = new Path(path)
     val fs = p.getFileSystem(configMapper.configuration)
+    if (!fs.exists(p)) {
+      return List[FileDescriptor]()
+    }
     fs.listStatus(p)
       .map((status: FileStatus) =>
         new FileDescriptor(status.getPath.getName, status.getModificationTime)
@@ -49,21 +50,20 @@ class ParallelDataLister extends DataLister with Serializable {
 
   override def listData(configMapper: ConfigMapper, path: Path): DataFrame = {
     import spark.implicits._
-    val fs = path.getFileSystem(spark.sparkContext.hadoopConfiguration)
-    val slices = listPath(configMapper, path.toString)
+    val slices = listPath(configMapper, path)
     val pathStr = path.toString
     val objectsUDF = udf((sliceId: String) => {
-      listPath(configMapper, pathStr + sliceId)
+      val slicePath = new Path(pathStr, sliceId)
+      listPath(configMapper, slicePath)
         .map(s => (s.path, s.lastModified))
     })
     val objectsDF = slices
       .map(_.path)
       .toDF("slice_id")
       .withColumn("udf", explode(objectsUDF(col("slice_id"))))
-      .withColumn("address", concat(col("slice_id"), lit("/"), col("udf._1")))
+      .withColumn("base_address", concat(col("slice_id"), lit("/"), col("udf._1")))
       .withColumn("last_modified", col("udf._2"))
-      .select("address", "last_modified")
-      .toDF("address", "last_modified")
+      .select("base_address", "last_modified")
     objectsDF
   }
 }