+++ pkg/block/mem/adapter.go
@@ -80,10 +80,6 @@ func getKey(obj block.ObjectPointer) string {
 	return fmt.Sprintf("%s:%s", obj.StorageNamespace, obj.Identifier)
 }
 
-func getPrefix(lsOpts block.WalkOpts) string {
-	return fmt.Sprintf("%s:%s", lsOpts.StorageNamespace, lsOpts.Prefix)
-}
-
 func (a *Adapter) Put(_ context.Context, obj block.ObjectPointer, _ int64, reader io.Reader, opts block.PutOpts) error {
 	a.mutex.Lock()
 	defer a.mutex.Unlock()
@@ -280,9 +276,15 @@ func (a *Adapter) Walk(_ context.Context, walkOpt block.WalkOpts, walkFn block.W
 	a.mutex.RLock()
 	defer a.mutex.RUnlock()
 
-	fullPrefix := getPrefix(walkOpt)
+	// Walk function should be performed in a lexicographical order
+	keys := make([]string, 0)
 	for k := range a.data {
-		if strings.HasPrefix(k, fullPrefix) {
+		keys = append(keys, k)
+	}
+	sort.Strings(keys)
+
+	for _, k := range keys {
+		if strings.HasPrefix(k, walkOpt.Prefix) {
 			if err := walkFn(k); err != nil {
 				return err
 			}